Simplify filter cmp structure.
